Got some more working
Everything in Main.java was working perfectly, then I un-commented the
line calling the static method in User (which, eventually, will call a
User constructor). It's something, again, with that damned Scanner
class, because the User.createUser static method also has a scanner,
and it works fine. The issue comes when we return from that method
back to Main.chooseProgram(), the next part of that method that asks
for user input results in an infinite loop.

Help me ...