Fixed some strange behavior
I had the program working mostly, but in the cases where, depending
on the user input, a function was called recursively, the exit
functionality looped the number of times the function was recursively
called before the program finally finished. So not an infinite loop,
but a bug nonetheless.

I fixed it by replacing all (both) the recursively called functions
with while (true) {}, break;, and return; control-flw functionality
Control-flow this way seems to be less buggy.