I don't even know what's happening anymore
I'm trying to build functionality that allows the user to create
instances of a class at their own discretion by putting a static
"factory" method in a while (true) loop. I created this new Student
class to play around with. Just trying to simplify to the point to
where I can grasp this particular concept.

I built a static method that returns a new instance of student with
the desired username. This method is called in the createUser method,
where the user provides the input for the username to allow the user.

And, not sure if I did this on this commit or the last, but I solved
all strange Scanner class behavior by just using one instance of the
Scanner class, declared in the Main class and closed in Main.main().
I just have to pass that instance as an argument to use it in other
classes/methods.