It's been a while
I keep forgetting to commit. My personal access token expired but I'm doing this from IntelliJ, so we'll see what happens.

I got a lot working since my last commit. I created the password complexity checker and a confirm password function within it. From the console, you can create as many users as you want, and then a for loop will display the details of each user when the program finishes.

The next thing I want to do is create a method somewhere that allows the user to specify which user account they want to interact with (like, for example, login to the account they choose). And I want to add security questions in case of a forgotten password.

It might be time to actually whip out the pen and paper, because, I need to figure out the basic design pattern of the creating users and logging into accounts you create, knowing that nothing is saved when the program terminates. Maybe I could somehow or another build this package into the chess package and have this entire project act as one unit rather than separate projects within a larger project.

Maybe I just wrote a bunch of gibberish, but regardless I need to start planning out these more complex projects on paper